Update AddProduct.kt & MyListProduct.kt @zeph1re

+++ app/src/main/java/and5/finalproject/secondhand5/view/fragment/seller/AddProduct.kt
@@ -67,12 +67,12 @@ class AddProduct : Fragment() {
     private fun addProduct() {
         add_btn.setOnClickListener {
             val productName = add_product_name.text.toString()
-            val productPrice = add_product_price.text.toString()
+            val productPrice = add_product_price.text.toString().toInt()
             val productDesc = add_product_desc.text.toString()
 
             val viewModelProduct = ViewModelProvider(this).get(ProductViewModel::class.java)
             userManager.userToken.asLiveData().observe(viewLifecycleOwner){
-                viewModelProduct.addSellerProduct(it,productName,productPrice.toInt(),productDesc)
+                viewModelProduct.addSellerProduct(it,productName,productPrice,productDesc)
                 Navigation.findNavController(requireView()).navigate(R.id.action_home_to_myListProduct)
             }
